Move sos Service and sos Server into pkg/sos
